requirements.txt
# Flask and Flask Extensions
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.25

# Forms and Validation
WTForms==3.1.2
email-validator==2.1.0.post1

# Environment and Configuration
python-dotenv==1.0.1

# HTTP Requests
requests==2.31.0

# Security
bcrypt==4.1.2

# Development Tools
Werkzeug==3.0.1
click==8.1.7
colorama==0.4.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5

# Utilities
alembic==1.13.1
blinker==1.7.0